Added realm methods and changed cell component
