+++ .idea/kotlinc.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="KotlinCommonCompilerArguments">
+    <option name="errors">
+      <ArgumentParseErrors />
+    </option>
+  </component>
+</project>
