+++ app/src/main/res/layout/activity_user_form.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/rootContainer"
     style="@style/RootContainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
