+++ app/src/main/res/values/strings_user_forms.xml
@@ -2,5 +2,6 @@
 <resources>
 
     <string name="user_forms_title">Adicionar usuário</string>
+    <string name="user_added_successfully">Usuário adicionado com sucesso!</string>
 
 </resources>
