+++ app/src/main/res/values/styles_containers.xml
@@ -15,7 +15,10 @@
     </style>
 
     <style name="CellContainer">
-        <item name="android:padding">@dimen/margin_medium</item>
+        <item name="android:paddingRight">@dimen/margin_medium</item>
+        <item name="android:paddingLeft">@dimen/margin_medium</item>
+        <item name="android:paddingTop">@dimen/margin_x_small</item>
+        <item name="android:paddingBottom">@dimen/margin_x_small</item>
         <item name="android:background">@android:color/white</item>
     </style>
