+++ app/src/main/res/values/styles_text.xml
@@ -7,15 +7,18 @@
     </style>
 
     <style name="BaseText.H2">
-        <item name="android:textSize">@dimen/text_size_large</item>
+        <item name="android:textSize">@dimen/text_size_x_large</item>
     </style>
 
     <style name="BaseText.H3">
-        <item name="android:textSize">@dimen/text_size_medium</item>
+        <item name="android:textSize">@dimen/text_size_large</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:layout_marginBottom">@dimen/margin_x_small</item>
     </style>
 
     <style name="BaseText.Body">
         <item name="android:textSize">@dimen/text_size_medium</item>
     </style>
+
 
 </resources>
